// packages/core/src/models/console-message.ts
export enum ConsoleMessage {
  TITLE = "Houston",
  BANNER = "Smart contract tool for cosmwasm",
  ERROR = "ERROR: ",
  SUCCESS = "SUCCESS: ",
  INFO = "INFO: ",
  GENERATE = "GENERATE: ",
  CREATE = "CREATE: ",
  UPDATE = "UPDATE: ",
  START_GENERATING = "Start generating contract boilerplate...",
}